refactoring: move dependencies check into separate file
